requirements.txt
fastavro==1.9.4
tensorflow-io==0.36.0
tensorflow==2.16.1 # this is the same version as the docker image, could also leave it unpinned if you just want to use the docker version
google-cloud-bigquery-storage==2.24.0